+++ src/states/controlState.ts
@@ -11,8 +11,6 @@ export interface ControlState {
     selectedServerId: string;
     selectedTime: Time | null;
     selectedTimeRange: TimeRange | null;
-    visibleTimeRange: TimeRange | null;
-    selectedFeatureId: string | null;
     selectedDrawMode: ol.geom.GeometryType | null;
     timeSeriesUpdateMode: 'add' | 'replace';
     timeAnimationActive: boolean;
@@ -33,8 +31,6 @@ export function newControlState(): ControlState {
         selectedServerId: VIEWER_DEFAULT_API_SERVER.id,
         selectedTime: null,
         selectedTimeRange: null,
-        visibleTimeRange: null,
-        selectedFeatureId: null,
         selectedDrawMode: 'Point',
         timeSeriesUpdateMode: 'add',
         timeAnimationActive: false,